All general functionality completed
The app now is fully usable though limited in its capabilities.

Separation of admin and public sections by use of namespaces for caching
purposes was done away with.  Instead now mod_rewrite with the use of
'admin' subdomain is used to determine if the request hits the framework
or uses cache.

Sweepers added to ensure stale caches are not served up.  The sweepers
try to work as efficiently as possible in the sense that they will not
clear a cache unless it is absolutely necessary.

General cleanup for files that should not have been tracked by git.